modified:   package-lock.json 	modified:   package.json 	renamed:    src/scripts/Container.jsx -> src/Components/Container.jsx 	renamed:    src/scripts/bar/Bar.jsx -> src/Components/bar/Bar.jsx 	renamed:    src/scripts/bar/BarContent.jsx -> src/Components/bar/BarContent.jsx 	renamed:    src/scripts/footer/Footer.jsx -> src/Components/footer/Footer.jsx 	renamed:    src/scripts/main/Main.jsx -> src/Components/main/Main.jsx 	new file:   src/Components/main/centerblock/CenterblockFilter.jsx 	new file:   src/Components/main/centerblock/CenterblockSearch.jsx 	new file:   src/Components/main/centerblock/MainCenterblock.jsx 	new file:   src/Components/main/centerblock/content/CenterblockContent.jsx 	new file:   src/Components/main/centerblock/content/ContentPlaylist.jsx 	new file:   src/Components/main/centerblock/content/ContentTitle.jsx 	new file:   src/Components/main/centerblock/content/PlaylistItem.jsx 	new file:   src/Components/main/centerblock/dropdown/CenterblockDropDown.jsx 	new file:   src/Components/main/centerblock/dropdown/DropdownItem.jsx 	new file:   src/Components/main/centerblock/dropdown/DropdownRadio.jsx 	new file:   src/Components/main/nav/MainNav.jsx 	new file:   src/Components/main/nav/NavBurger.jsx 	new file:   src/Components/main/nav/NavLogo.jsx 	renamed:    src/scripts/main/MainNav.jsx -> src/Components/main/nav/NavMenu.jsx 	new file:   src/Components/main/sidebar/MainSidebar.jsx 	new file:   src/Components/main/sidebar/SidebarBlock.jsx 	new file:   src/Components/main/sidebar/SidebarItem.jsx 	new file:   src/Components/main/sidebar/SidebarPersonal.jsx 	renamed:    src/scripts/reportWebVitals.js -> src/Components/reportWebVitals.js 	renamed:    src/scripts/setupTests.js -> src/Components/setupTests.js 	new file:   src/assets/icon/ellipse.svg 	new file:   src/assets/icon/filledellipse.svg 	new file:   src/css/_dropdown.scss 	modified:   src/css/index.css 	modified:   src/css/index.css.map 	modified:   src/css/index.scss 	modified:   src/index.js
