+++ src/Components/main/Main.jsx
@@ -1,6 +1,6 @@
-import { MainNav } from './MainNav';
-import { MainCenterblock } from './MainCenterblock';
-import { MainSidebar } from './MainSidebar';
+import { MainNav } from './nav/MainNav';
+import { MainCenterblock } from './centerblock/MainCenterblock';
+import { MainSidebar } from './sidebar/MainSidebar';
 
 export const Main = () => {
     return (
