+++ src/Components/main/centerblock/CenterblockSearch.jsx
@@ -1,0 +1,15 @@
+export const CenterblockSearch = () => {
+    return (
+        <div className="centerblock__search search">
+            <svg className="search__svg">
+                <use xlinkHref="img/icon/sprite.svg#icon-search"></use>
+            </svg>
+            <input
+                className="search__text"
+                type="search"
+                placeholder="Поиск"
+                name="search"
+            ></input>
+        </div>
+    );
+}
