+++ src/Components/main/centerblock/dropdown/DropdownItem.jsx
@@ -1,0 +1,7 @@
+export const DropdownItem = (props) => {
+    return (
+        <li className="dropdownItem">
+            <span> {props.text} </span>
+        </li>
+    );
+};
