+++ src/Components/main/nav/NavBurger.jsx
@@ -1,0 +1,13 @@
+export const NavBurger = ({ onClick }) => {
+    const changeVisibility = () => {
+        onClick();
+    };
+
+    return (
+        <div onClick={changeVisibility} className="nav__burger burger">
+            <span className="burger__line"></span>
+            <span className="burger__line"></span>
+            <span className="burger__line"></span>
+        </div>
+    );
+}
