+++ src/Components/main/sidebar/MainSidebar.jsx
@@ -1,0 +1,17 @@
+import { SidebarPersonal } from "./SidebarPersonal";
+import { SidebarBlock } from "./SidebarBlock";
+
+export const MainSidebar = () => {
+    return (
+        <div className="main__sidebar sidebar">
+            <SidebarPersonal name="Sergey.Ivanov" />
+            <SidebarBlock />
+        </div>
+    );
+};
+
+
+
+
+
+
