+++ src/Components/main/sidebar/SidebarItem.jsx
@@ -1,0 +1,13 @@
+export const SidebarItem = (props) => {
+    return (
+        <div className="sidebar__item">
+            <a className="sidebar__link" href="#">
+                <img
+                    className="sidebar__img"
+                    src={props.src}
+                    alt="day's playlist"
+                ></img>
+            </a>
+        </div>
+    );
+}
