+++ src/css/index.css
@@ -1,3 +1,4 @@
+@charset "UTF-8";
 * {
   margin: 0;
   padding: 0;
@@ -27,7 +28,7 @@ ul li {
 
 @font-face {
   font-family: "StratosSkyeng";
-  src: local("StratosSkyeng"), local("StratosSkyeng"), url("../assets/fonts/StratosSkyeng/Stratos-Black.woff2") format("woff2"), url("../assets/fonts/StratosSkyeng/Stratos-Black.woff") format("woff");
+  src: local("StratosSkyeng"), local("StratosSkyeng"), url("../assets/fonts/StratosSkyeng/Stratos-Regular.woff2") format("woff2"), url("../assets/fonts/StratosSkyeng/Stratos-Regular.woff") format("woff");
   font-weight: 400;
   font-style: normal;
 }
@@ -88,7 +89,6 @@ body {
 .main {
   flex: 1 1 auto;
   display: flex;
-  flex-wrap: wrap;
   justify-content: space-between;
 }
 .main__nav {
@@ -97,13 +97,13 @@ body {
   padding: 20px 0 20px 36px;
 }
 .main__centerblock {
-  width: auto;
+  width: 70vw;
   flex-grow: 3;
-  padding: 20px 40px 20px 111px;
+  padding: 20px 40px 20px 0px;
 }
 .main__sidebar {
   max-width: 418px;
-  padding: 20px 90px 20px 78px;
+  padding: 20px 90px 20px 0px;
 }
 
 .nav__logo {
@@ -178,6 +178,7 @@ body {
 .centerblock__content {
   display: flex;
   flex-direction: column;
+  margin-right: 178px;
 }
 
 .search__svg {
@@ -238,6 +239,12 @@ body {
   border: 1px solid #ffffff;
   border-radius: 60px;
   padding: 6px 20px;
+  transition: 0.2s all;
+}
+.filter__button:focus {
+  border-color: #ad61ff;
+  color: #ad61ff;
+  cursor: pointer;
 }
 .filter__button:not(:last-child) {
   margin-right: 10px;
@@ -273,7 +280,7 @@ body {
 }
 
 .col01 {
-  width: 447px;
+  width: 230px;
 }
 
 .col02 {
@@ -305,7 +312,7 @@ body {
   display: flex;
   flex-direction: row;
   align-items: center;
-  width: 447px;
+  min-width: 230px;
 }
 .track__title-image {
   width: 51px;
@@ -612,4 +619,141 @@ body {
 }
 .volume__progress-line {
   width: 109px;
+}
+
+.dropdown-div {
+  width: 300px;
+  height: 200px;
+}
+
+.dropdown-menu {
+  position: absolute;
+  min-width: 248px;
+  max-height: 305px;
+  background: #313131;
+  border-radius: 12px;
+  padding-top: 36.5px;
+  padding-left: 34px;
+  padding-right: 34px;
+  padding-bottom: 36.5px;
+  box-sizing: border-box;
+}
+.dropdown-menu::before {
+  content: "";
+  position: absolute;
+  top: -5px;
+  right: 20px;
+  height: 20px;
+  width: 20px;
+  background: var(--secondary-bg);
+  transform: rotate(45deg);
+}
+
+.dropdown-menu.active {
+  opacity: 1;
+  visibility: visible;
+  transform: translateY(0);
+  transition: var(--speed) ease;
+}
+
+.dropdown-menu.inactive {
+  opacity: 0;
+  visibility: hidden;
+  transform: translateY(-20px);
+  transition: var(--speed) ease;
+}
+
+.dropdown-menu li:hover span {
+  cursor: pointer;
+}
+
+.dropdown-menu ul {
+  width: 176px;
+  max-height: 232px;
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 28px;
+  box-sizing: border-box;
+}
+
+.dropdown-menu ul::-webkit-scrollbar {
+  width: 4px; /* ширина всей полосы прокрутки */
+}
+
+.dropdown-menu ul::-webkit-scrollbar-track {
+  background: #4b4949;
+  border-radius: 10px; /* цвет зоны отслеживания */
+  height: 65px;
+}
+
+.dropdown-menu ul::-webkit-scrollbar-thumb {
+  background: #ffffff;
+  border-radius: 10px;
+}
+
+.dropdownItem {
+  display: flex;
+}
+
+.dropdownItem span {
+  width: 119px;
+  height: 24px;
+  font-family: "StratosSkyeng";
+  font-style: normal;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 24px;
+  transition: 0.2s all;
+}
+.dropdownItem span:hover {
+  color: #ad61ff;
+  text-decoration: underline;
+}
+.dropdownItem span:active {
+  color: #ad61ff;
+  text-decoration: underline;
+}
+
+.dropdown__radio-div {
+  min-width: 300px;
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 28px;
+  text-align: center;
+}
+
+.dropdown__radio-div input {
+  display: none;
+}
+
+.dropdown__radio-div label {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 24px;
+  text-align: center;
+  cursor: pointer;
+}
+
+.dropdown__radio-div label::before {
+  text-align: center;
+  width: 18px;
+  height: 18px;
+  content: "";
+  margin-right: 10px;
+  display: inline-block;
+  background: url("../assets/icon/ellipse.svg");
+  cursor: pointer;
+}
+
+.dropdown__radio-div input:checked + label::before {
+  margin-right: 10px;
+  width: 18px;
+  height: 18px;
+  content: "";
+  display: inline-block;
+  background: url("../assets/icon/filledellipse.svg");
 }/*# sourceMappingURL=index.css.map */
