+++ src/css/index.scss
@@ -31,8 +31,9 @@ ul li {
 @font-face {
     font-family: 'StratosSkyeng';
     src: local('StratosSkyeng'), local('StratosSkyeng'),
-        url('../assets/fonts/StratosSkyeng/Stratos-Black.woff2') format('woff2'),
-        url('../assets/fonts/StratosSkyeng/Stratos-Black.woff') format('woff');
+        url('../assets/fonts/StratosSkyeng/Stratos-Regular.woff2')
+            format('woff2'),
+        url('../assets/fonts/StratosSkyeng/Stratos-Regular.woff') format('woff');
     font-weight: 400;
     font-style: normal;
 }
@@ -100,7 +101,7 @@ body {
 .main {
     flex: 1 1 auto;
     display: flex;
-    flex-wrap: wrap;
+    // flex-wrap: wrap;
     justify-content: space-between;
 
     &__nav {
@@ -109,15 +110,16 @@ body {
         padding: 20px 0 20px 36px;
     }
 
+    // изменила пэддинг с 111px на 0px и width с auto на 70vw
     &__centerblock {
-        width: auto;
+        width: 70vw;
         flex-grow: 3;
-        padding: 20px 40px 20px 111px;
+        padding: 20px 40px 20px 0px;
     }
 
     &__sidebar {
         max-width: 418px;
-        padding: 20px 90px 20px 78px;
+        padding: 20px 90px 20px 0px;
     }
 }
 
@@ -209,6 +211,7 @@ body {
     &__content {
         display: flex;
         flex-direction: column;
+        margin-right: 178px;
     }
 }
 
@@ -261,6 +264,13 @@ body {
         border-radius: 60px;
         padding: 6px 20px;
 
+        transition: 0.2s all;
+
+        &:focus {
+            border-color: #ad61ff;
+            color: #ad61ff;
+            cursor: pointer;
+        }
         &:not(:last-child) {
             margin-right: 10px;
         }
@@ -303,7 +313,7 @@ body {
 }
 
 .col01 {
-    width: 447px;
+    width: 230px;
 }
 
 .col02 {
@@ -338,7 +348,8 @@ body {
         display: flex;
         flex-direction: row;
         align-items: center;
-        width: 447px;
+        // width: 447px;
+        min-width: 230px;
     }
 
     &__title-image {
@@ -717,3 +728,6 @@ body {
         width: 109px;
     }
 }
+
+
+@import 'dropdown';
