+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './css/index.css';
-import reportWebVitals from './scripts/reportWebVitals';
-import { Container } from './scripts/Container';
+import reportWebVitals from './Components/reportWebVitals';
+import { Container } from './Components/Container';
 
 const wrapper = ReactDOM.createRoot(document.querySelector('.wrapper'));
 wrapper.render(
